feat: set up provider edit page
