+++ client/src/app/[tenant]/agents/[taskId]/[taskSchemaId]/deployments/DeploymentsContainer.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import { useEnvDeploy } from '@/components/DeployIterationModal/DeployVersionModal';
-import { AddProviderKeyModal } from '@/components/ProviderKeysModal/AddProviderKeyModal';
-import { useFindProviderConfigID } from '@/components/ProviderKeysModal/useFindProviderConfigID';
 import { Button } from '@/components/ui/Button';
 import { PageContainer } from '@/components/v2/PageContainer';
 import { PROVIDER_KEYS_MODAL_OPEN, useQueryParamModal } from '@/lib/globalModal';
@@ -14,10 +12,9 @@ import { useOrFetchOrganizationSettings, useOrFetchTask, useOrFetchVersions } fr
 import { useOrganizationSettings } from '@/store/organization_settings';
 import { useVersions } from '@/store/versions';
 import { TaskSchemaID } from '@/types/aliases';
-import { Model, VersionV1 } from '@/types/workflowAI';
+import { VersionV1 } from '@/types/workflowAI';
 import { DeployVersionModal, EditEnvSchemaIterationParams } from './DeployVersionModal';
 import { EnvironmentDeployment } from './EnvironmentDeployment';
-import { UpdateProviderModal } from './UpdateProviderModal';
 
 export function DeploymentsContainer() {
   const { tenant, taskId, taskSchemaId } = useTaskSchemaParams();
@@ -25,7 +22,7 @@ export function DeploymentsContainer() {
 
   const {
     versions,
-    deployedVersions,
+
     versionsPerEnvironment,
     isInitialized: isVersionsInitialized,
   } = useOrFetchVersions(tenant, taskId);
@@ -35,77 +32,6 @@ export function DeploymentsContainer() {
   const [envSchemaIteration, setEnvSchemaIteration] = useState<EditEnvSchemaIterationParams | undefined>();
 
   const onCloseEnvSchemaIteration = useCallback(() => setEnvSchemaIteration(undefined), [setEnvSchemaIteration]);
-
-  const [selectedVersion, setSelectedVersion] = useState<VersionV1 | undefined>();
-
-  const onCloseUpdateProviderModal = useCallback(() => setSelectedVersion(undefined), [setSelectedVersion]);
-
-  const [addProviderKeyModalOpen, setAddProviderKeyModalOpen] = useState(false);
-  const onCloseAddProviderKeyModal = useCallback(() => setAddProviderKeyModalOpen(false), [setAddProviderKeyModalOpen]);
-
-  const { organizationSettings, isInitialized: isOrganizationSettingsInitialized } = useOrFetchOrganizationSettings();
-
-  const addProviderConfig = useOrganizationSettings((state) => state.addProviderConfig);
-
-  const model: Model | string | undefined = selectedVersion?.model;
-
-  const { providerConfigID: settingsProviderConfigID } = useFindProviderConfigID({
-    tenant,
-    taskId,
-    taskSchemaId,
-    providerSettings: organizationSettings?.providers,
-    model: model,
-  });
-
-  const deployVersion = useVersions((state) => state.deployVersion);
-  const environment = selectedVersion?.deployments?.[0]?.environment;
-
-  const onToggleProviderKey = useCallback(
-    async (useWorkflowAIKey: boolean) => {
-      if (!selectedVersion?.schema_id || !selectedVersion.iteration || !environment) {
-        return;
-      }
-
-      if (!useWorkflowAIKey && !settingsProviderConfigID) {
-        setAddProviderKeyModalOpen(true);
-        return;
-      }
-
-      const payload = {
-        environment,
-        provider_config_id: useWorkflowAIKey ? undefined : settingsProviderConfigID,
-      };
-
-      const newSelectedVersion = await deployVersion(
-        tenant,
-        taskId,
-        selectedVersion.schema_id?.toString() as TaskSchemaID,
-        selectedVersion.id,
-        selectedVersion.iteration,
-        payload
-      );
-
-      setSelectedVersion(newSelectedVersion);
-    },
-    [deployVersion, tenant, taskId, settingsProviderConfigID, environment, selectedVersion]
-  );
-
-  const [shouldToggleProviderKey, setShouldToggleProviderKey] = useState(false);
-  const handleAddProviderKeyModalClose = useCallback(
-    (keyWasAdded: boolean) => {
-      if (keyWasAdded) {
-        setShouldToggleProviderKey(true);
-      }
-      onCloseAddProviderKeyModal();
-    },
-    [onCloseAddProviderKeyModal]
-  );
-  useEffect(() => {
-    if (shouldToggleProviderKey && !!settingsProviderConfigID) {
-      onToggleProviderKey(false);
-      setShouldToggleProviderKey(false);
-    }
-  }, [shouldToggleProviderKey, onToggleProviderKey, settingsProviderConfigID]);
 
   const favoriteVersions = useMemo(() => versions.filter((version) => version.is_favorite), [versions]);
 
@@ -129,31 +55,11 @@ export function DeploymentsContainer() {
       return;
     }
 
-    const providerConfigId = !!envSchemaIteration.currentIteration
-      ? deployedVersions.find((version) => version?.iteration?.toString() === envSchemaIteration.iteration?.toString())
-          ?.deployments?.[0]?.provider_config_id
-      : undefined;
-
-    await deployVersionToEnv(
-      envSchemaIteration.environment,
-      versionId,
-      versionText,
-      envSchemaIteration?.iteration,
-      providerConfigId
-    );
+    await deployVersionToEnv(envSchemaIteration.environment, versionId, versionText);
 
     await fetchVersions(tenant, taskId, undefined);
     onCloseEnvSchemaIteration();
-  }, [
-    deployVersionToEnv,
-    envSchemaIteration,
-    fetchVersions,
-    tenant,
-    taskId,
-    onCloseEnvSchemaIteration,
-    deployedVersions,
-    versions,
-  ]);
+  }, [deployVersionToEnv, envSchemaIteration, fetchVersions, tenant, taskId, onCloseEnvSchemaIteration, versions]);
 
   const onIterationChange = useCallback(
     (iteration: string) => {
@@ -177,7 +83,6 @@ export function DeploymentsContainer() {
     tenant,
     taskId,
     setEnvSchemaIteration,
-    setSelectedVersion,
     taskSchemaId,
     versionsPerEnvironment,
     isInDemoMode,
@@ -188,7 +93,7 @@ export function DeploymentsContainer() {
       name='Deployments'
       task={task}
       showSchema={false}
-      isInitialized={isTaskInitialized && isVersionsInitialized && isOrganizationSettingsInitialized}
+      isInitialized={isTaskInitialized && isVersionsInitialized}
       documentationLink='https://docs.workflowai.com/features/deployments'
       rightBarChildren={
         <div className='flex flex-row items-center gap-2 font-lato'>
@@ -218,18 +123,6 @@ export function DeploymentsContainer() {
         allVersions={versions}
         tenant={tenant}
         taskId={taskId}
-      />
-      <UpdateProviderModal
-        selectedVersion={selectedVersion}
-        onClose={onCloseUpdateProviderModal}
-        onToggleProviderKey={onToggleProviderKey}
-      />
-      <AddProviderKeyModal
-        open={addProviderKeyModalOpen}
-        onClose={handleAddProviderKeyModalClose}
-        currentProvider={selectedVersion?.properties?.provider}
-        organizationSettings={organizationSettings}
-        addProviderConfig={addProviderConfig}
       />
     </PageContainer>
   );
